// frontend/src/types/auth.types.ts
export type RegisterFormValues = {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
};

export type RegisterDTO = {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
};

export type LoginDTO = {
  email: string;
  password: string;
};

export type LoginFormValues = {
  email: string;
  password: string;
};